Renamed thunk TS files to asyncActionCreator.ts
