+++ src/redux/callState/actionCreator.ts
@@ -1,6 +1,5 @@
-import { OutcomeData } from './callThunk';
 import { Issue } from '../../common/model';
-import { CurrentIssueAction, SubmitOutcomeAction, NextContact } from './index';
+import { OutcomeData, CurrentIssueAction, SubmitOutcomeAction, NextContact } from './index';
 
 export const selectIssueActionCreator = (issue: Issue): CurrentIssueAction => {
   return {
