+++ src/redux/callState/index.ts
@@ -4,4 +4,4 @@ export { CallStateAction, CallStateActionType,
   SubmitOutcomeAction } from './action';
 export {completeIssueActionCreator, moveToNextActionCreator, selectIssueActionCreator} from './actionCreator';
 export { callStateReducer, CallState } from './reducer';
-export { OutcomeData, OutcomeType, submitOutcome } from './callThunk';
+export { OutcomeData, OutcomeType, submitOutcome } from './asyncActionCreator';
