+++ src/redux/location/index.ts
@@ -1,4 +1,4 @@
 export { LocationAction, LocationActionType, LocationClearedAction, LocationSetAction } from './action';
 export { setLocation, clearAddress } from './actionCreator';
 export { LocationState, locationStateReducer } from './reducer';
-export { setAddress } from './thunk';
+export { setAddress } from './asyncActionCreator';
