+++ src/redux/remoteData/index.ts
@@ -1,4 +1,4 @@
 export { IssuesAction, RemoteDataAction, RemoteDataActionType, CallCountAction } from './action';
 export { issuesActionCreator, callCountActionCreator } from './actionCreator';
 export { RemoteDataState, remoteDataReducer } from './reducer';
-export { getIssues, fetchCallCount } from './apiThunk';
+export { getIssues, fetchCallCount } from './asyncActionCreator';
